Using memory rate limiter for authorization server
