+++ packages/agreement-email-sender/test/utils.ts
@@ -33,7 +33,6 @@ export const {
   undefined,
   undefined,
   emailManagerConfig,
-  undefined,
   sesEmailManagerConfig
 );
 export const readModelService = readModelServiceBuilder(readModelRepository);
