+++ packages/commons-test/test/memoryRateLimiter.integration.test.ts
@@ -1,20 +1,16 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { describe, expect, it } from "vitest";
 import { generateId, TenantId } from "pagopa-interop-models";
-import { genericLogger, initRedisRateLimiter } from "pagopa-interop-commons";
+import { genericLogger, initMemoryRateLimiter } from "pagopa-interop-commons";
 import { sleep } from "../src/testUtils.js";
-import { redisRateLimiterConfig } from "./utils.js";
+import { memoryRateLimiterConfig } from "./utils.js";
 
-const redisRateLimiter = await initRedisRateLimiter({
+const memoryRateLimiter = await initMemoryRateLimiter({
   limiterGroup: "TEST",
-  maxRequests: redisRateLimiterConfig.rateLimiterMaxRequests,
-  rateInterval: redisRateLimiterConfig.rateLimiterRateInterval,
-  burstPercentage: redisRateLimiterConfig.rateLimiterBurstPercentage,
-  redisHost: redisRateLimiterConfig.rateLimiterRedisHost,
-  redisPort: redisRateLimiterConfig.rateLimiterRedisPort,
-  timeout: redisRateLimiterConfig.rateLimiterTimeout,
+  maxRequests: memoryRateLimiterConfig.rateLimiterMaxRequests,
+  rateInterval: memoryRateLimiterConfig.rateLimiterRateInterval,
+  burstPercentage: memoryRateLimiterConfig.rateLimiterBurstPercentage,
 });
-
 describe("Redis rate limiter tests", async () => {
   /*
   ---------- NOTE ------------------------------------
@@ -25,12 +21,8 @@ describe("Redis rate limiter tests", async () => {
   it("should rate limit requests by organizationId", async () => {
     const organizationId: TenantId = generateId();
 
-    expect(await redisRateLimiter.getCountByOrganization(organizationId)).toBe(
-      0
-    );
-
     expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -41,12 +33,8 @@ describe("Redis rate limiter tests", async () => {
       remainingRequests: 1,
     });
 
-    expect(await redisRateLimiter.getCountByOrganization(organizationId)).toBe(
-      1
-    );
-
     expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -56,16 +44,12 @@ describe("Redis rate limiter tests", async () => {
       rateInterval: 1000,
       remainingRequests: 0,
     });
-
-    expect(await redisRateLimiter.getCountByOrganization(organizationId)).toBe(
-      2
-    );
 
     // Burst rate limiter kicks in.
     // Burst percentage in config is 1.5, so we expect 3 requests to be allowed.
 
     expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -77,11 +61,7 @@ describe("Redis rate limiter tests", async () => {
     });
 
     expect(
-      await redisRateLimiter.getBurstCountByOrganization(organizationId)
-    ).toBe(1);
-
-    expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -93,11 +73,7 @@ describe("Redis rate limiter tests", async () => {
     });
 
     expect(
-      await redisRateLimiter.getBurstCountByOrganization(organizationId)
-    ).toBe(2);
-
-    expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -107,14 +83,10 @@ describe("Redis rate limiter tests", async () => {
       rateInterval: 1000,
       remainingRequests: 0,
     });
-
-    expect(
-      await redisRateLimiter.getBurstCountByOrganization(organizationId)
-    ).toBe(3);
 
     // Limit reached
     expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -129,12 +101,8 @@ describe("Redis rate limiter tests", async () => {
   it("should reset requests count after rate interval", async () => {
     const organizationId: TenantId = generateId();
 
-    expect(await redisRateLimiter.getCountByOrganization(organizationId)).toBe(
-      0
-    );
-
     expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -144,15 +112,11 @@ describe("Redis rate limiter tests", async () => {
       rateInterval: 1000,
       remainingRequests: 1,
     });
-
-    expect(await redisRateLimiter.getCountByOrganization(organizationId)).toBe(
-      1
-    );
 
     await sleep(1000);
 
     expect(
-      await redisRateLimiter.rateLimitByOrganization(
+      await memoryRateLimiter.rateLimitByOrganization(
         organizationId,
         genericLogger
       )
@@ -162,9 +126,5 @@ describe("Redis rate limiter tests", async () => {
       rateInterval: 1000,
       remainingRequests: 1,
     });
-
-    expect(await redisRateLimiter.getCountByOrganization(organizationId)).toBe(
-      1
-    );
   });
 });
