+++ packages/commons/src/config/index.ts
@@ -12,7 +12,7 @@ export * from "./fileManagerConfig.js";
 export * from "./kafkaTopicConfig.js";
 export * from "./tokenGenerationConfig.js";
 export * from "./sessionTokenGenerationConfig.js";
-export * from "./redisRateLimiterConfig.js";
+export * from "./rateLimiterConfig.js";
 export * from "./pecEmailManagerConfig.js";
 export * from "./selfcareConfig.js";
 export * from "./featureFlagsConfig.js";
