+++ packages/commons/src/rate-limiter/index.ts
@@ -1,4 +1,4 @@
 export * from "./rateLimiterMiddleware.js";
 export * from "./rateLimiterModel.js";
 export * from "./rateLimiterUtils.js";
-export * from "./redisRateLimiter.js";
+export * from "./rateLimiter.js";
