+++ packages/commons/src/rate-limiter/rateLimiterMiddleware.ts
@@ -6,13 +6,13 @@ import {
   tooManyRequestsError,
 } from "pagopa-interop-models";
 import { ExpressContext, fromAppContext } from "../context/context.js";
-import { RateLimiter } from "./rateLimiterModel.js";
+import { RedisRateLimiter } from "./rateLimiterModel.js";
 import { rateLimiterHeadersFromStatus } from "./rateLimiterUtils.js";
 
 const makeApiProblem = makeApiProblemBuilder({});
 
 export function rateLimiterMiddleware(
-  rateLimiter: RateLimiter
+  rateLimiter: RedisRateLimiter
 ): ZodiosRouterContextRequestHandler<ExpressContext> {
   return async (req, res, next) => {
     const ctx = fromAppContext(req.ctx);
